Improved tests to be more isolated
